+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/update/MediaGroupUpdates/EditMessageMediaGroupUpdate.kt
@@ -1,13 +1,4 @@
 package dev.inmo.tgbotapi.types.update.MediaGroupUpdates
 
-import dev.inmo.tgbotapi.types.UpdateIdentifier
-import dev.inmo.tgbotapi.types.message.abstracts.MediaGroupMessage
-import dev.inmo.tgbotapi.types.message.content.MediaGroupContent
-import dev.inmo.tgbotapi.types.update.EditMessageUpdate
-
-data class EditMessageMediaGroupUpdate(
-    override val origin: EditMessageUpdate
-) : EditMediaGroupUpdate {
-    override val updateId: UpdateIdentifier = origin.updateId
-    override val data: MediaGroupMessage<MediaGroupContent> = origin.data as MediaGroupMessage<MediaGroupContent>
-}
+@Deprecated("Replaced", ReplaceWith("EditMessageMediaGroupUpdate", "dev.inmo.tgbotapi.types.update.media_group.EditMessageMediaGroupUpdate"))
+typealias EditMessageMediaGroupUpdate = dev.inmo.tgbotapi.types.update.media_group.EditMessageMediaGroupUpdate