+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/updateshandlers/FlowsUpdatesFilter.kt
@@ -3,9 +3,9 @@ package dev.inmo.tgbotapi.updateshandlers
 import dev.inmo.micro_utils.coroutines.plus
 import dev.inmo.tgbotapi.types.ALL_UPDATES_LIST
 import dev.inmo.tgbotapi.types.update.*
-import dev.inmo.tgbotapi.types.update.MediaGroupUpdates.*
 import dev.inmo.tgbotapi.types.update.abstracts.UnknownUpdate
 import dev.inmo.tgbotapi.types.update.abstracts.Update
+import dev.inmo.tgbotapi.types.update.media_group.*
 import kotlinx.coroutines.channels.*
 import kotlinx.coroutines.flow.*